fix: Fixed geoHeadingAccuracy
Moved all geopose accuracy metadata for the “user” entity into the “stage” entity. This is currently a bit confusing, since this metadata describes the accuracy of the user pose relative to FIXED (not necessarily the accuracy of the stage pose or any other geopositioned entity).  However, if a reality overrides the stage pose, then it should be able to redefine the accuracy metadata without also needing to override the user entity. To make it less confusing, we may want to consider renaming this particular metadata property to something like “userHeadingAccuracy” to make it clear that it refers to the “user” entity.